requirements.txt
pytest>=7.0
python-dotenv>=1.0
streamlit>=1.35

# Optional manual-demo extras
# Install local embedder support with:
#   pip install sentence-transformers
# Install OpenAI embedder support with:
#   pip install openai
# Install both optional embedders with:
#   pip install openai sentence-transformers
# sentence-transformers>=3.0
# openai>=1.0

# Optional: for bonus tasks
chromadb>=0.4.0      # Real vector database integration
